refactor: move code to components
